langchain/pyproject.toml
[tool.poetry]
name = "Learning langchain"
version="0.1.0"
authors = ["Jay G"]
description = "Learning langchain"
[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
langchain="^0.2.1"
langchain-groq = "0.1.9"
langchain-community="^0.2.1"
langchain-google-genai="^1.0.5"
langchain-google-firestore="^0.3.0"
python-dotenv = "^1.0.1"
langchainhub="^0.1.18"
firecrawl-py="^0.0.13"
tiktoken="^0.7.0"
wikipedia = "^1.4.0"
chromadb="^0.5.0"
bs4="^0.0.2"
firestore="^0.0.8"
sentence-transformers="^3.0.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes=true
exclude = [".cache"]

